Moved /clear-corpses command to commands dir; Added default radius to /clear-corpses; Tweaked naming in Functions.clear_corpses; Replaced errors with warning/infos in /clear_corpses
