+++ commands.lua
@@ -5,3 +5,4 @@ require('commands.inventory_costs')
 require('commands.current_map_seed')
 require('commands.suspend')
 require('commands.where')
+require('commands.clear_corpses')
